update: changed the file structure to use dtos and mappers as well as service interface and implementation...need to resolve error from merge
