+++ backend/src/main/java/com/algoExpert/demo/Service/CommentService.java
@@ -22,72 +22,21 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class CommentService {
-
-    @Autowired
-    private CommentRepository commentRepository;
-    @Autowired
-    private UserRepository userRepository ;
-    @Autowired
-    private TaskRepository taskRepository;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private CommentMapper commentMapper;
-    @Autowired
-    private TaskMapper taskMapper;
 public interface CommentService {
 
-//    create comment
-    public TaskDto createComment(CommentDto commentDto, int member_id, int task_id)throws InvalidArgument {
-        Comment comment = commentMapper.commentDtoToComment(commentDto);
-        Member findMember = memberRepository.findById(member_id).orElseThrow(() -> new InvalidArgument("Task with ID " + member_id + " not found"));
-        Task task = taskRepository.findById(task_id).orElseThrow(() -> new InvalidArgument("Task with ID " + task_id + " not found"));
-        User user = userRepository.findById(findMember.getUser_id()).get();
-        comment.setUsername(user.getUsername());
+    //    create comment
+    TaskDto createComment(CommentDto commentDto, int member_id, int task_id) throws InvalidArgument;
 
-        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd/MMM/yyyy HH:mm:ss");
-        comment.setDate_created(simpleDateFormat.format(new Date()));
-        List<Comment> commentList = task.getComments();
-        commentList.add(comment);
+    //  create comment
+    Task createComment(Comment comment, int member_id, int task_id) throws InvalidArgument;
 
-        task.setComments(commentList);
-        Task taskResult = taskRepository.save(task);
-        return taskMapper.taskToTaskDto(taskResult);
-    }
-//  create comment
-    Task createComment(Comment comment,int member_id,int task_id)throws InvalidArgument;
+    //    get all comments
+    public List<CommentDto> getAllComments();
 
-//    get all comments
-    public List<CommentDto> getAllComments(){
-    List<Comment> comments = commentRepository.findAll();
-    return commentMapper.commentDtos(comments);
-}
-//  get all comments
-    List<Comment> getAllComments();
+    // update comment by id
+    public CommentDto editComment(int commentId, CommentDto newCommentDto) throws InvalidArgument;
 
-// update comment by id
-    public CommentDto editComment(int commentId,CommentDto newCommentDto) throws InvalidArgument{
-        Comment comment= commentRepository.findById(commentId)
-                .map(oldComment -> {
-                    if(newCommentDto != null){
-                        Optional.ofNullable(newCommentDto.getCommentBody()).ifPresent(oldComment::setCommentBody);
-                    }
-                return commentRepository.save(oldComment);}).orElseThrow(()-> new InvalidArgument("Comment ID "+commentId+" not fount"));
-        return commentMapper.commentToCommentDto(comment);
-    }
-//   update comment by id
-    Comment editComment(int commentId,Comment newComment);
-
-//    delete comment by id
-    public List<Comment> deleteComment(int commentId){
-        Comment findComment = commentRepository.findById(commentId).orElseThrow(()->new IllegalArgumentException("comment with Id "+commentId+" does not exist"));
-        if(findComment !=null){
-            commentRepository.deleteById(commentId);
-            return commentRepository.findAll();
-        }
-
-        return null;
-    }
+    //    delete comment by id
+    public List<Comment> deleteComment(int commentId);
 
 }
