+++ backend/src/main/java/com/algoExpert/demo/Service/Impl/MemberServiceImpl.java
@@ -27,15 +27,15 @@ public class MemberServiceImpl implements MemberService {
     @Autowired
     private MemberRepository memberRepository;
 
-//    Invite member to project
+    //    Invite member to project
     @Override
-    public Member inviteMember (int project_id , int user_id)throws InvalidArgument{
+    public Member inviteMember(int project_id, int user_id) throws InvalidArgument {
         // check if user and project exist
-        User user = userRepository.findById(user_id).orElseThrow(()->
-                new InvalidArgument("User wth ID "+user_id+" not found"));
+        User user = userRepository.findById(user_id).orElseThrow(() ->
+                new InvalidArgument("User wth ID " + user_id + " not found"));
 
-        Project userProject = projectRepository.findById(project_id).orElseThrow(()->
-                new InvalidArgument("Project wth ID "+project_id+" not found"));
+        Project userProject = projectRepository.findById(project_id).orElseThrow(() ->
+                new InvalidArgument("Project wth ID " + project_id + " not found"));
 
         // Initialize the members list if it's null
         List<Member> members = userProject.getMemberList();
@@ -46,13 +46,13 @@ public class MemberServiceImpl implements MemberService {
         // check if member exist
         boolean memberExist = members.stream()
                 .map(Member::getUser_id)
-                .anyMatch(id->id==user_id);
+                .anyMatch(id -> id == user_id);
 
-        if(memberExist){
+        if (memberExist) {
             throw new InvalidArgument("User ID " + user_id + " is already a member");
-        }else{
+        } else {
             // create a new member
-            Member newMember = new Member(0,user.getUser_id(), userProject.getProject_id(),null);
+            Member newMember = new Member(0, user.getUser_id(), userProject.getProject_id(), null);
             members.add(newMember);
             userProject.setMemberList(members);
             projectRepository.save(userProject);
@@ -61,9 +61,9 @@ public class MemberServiceImpl implements MemberService {
         }
     }
 
-//    get all members
+    //    get all members
     @Override
-    public List<Member> getAllMembers(){
+    public List<Member> getAllMembers() {
         return memberRepository.findAll();
     }
