+++ backend/src/main/java/com/algoExpert/demo/Service/Impl/ProjectServiceImpl.java
@@ -1,7 +1,9 @@
 package com.algoExpert.demo.Service.Impl;
 
+import com.algoExpert.demo.Dto.ProjectDto;
 import com.algoExpert.demo.Entity.*;
 import com.algoExpert.demo.ExceptionHandler.InvalidArgument;
+import com.algoExpert.demo.Mapper.ProjectMapper;
 import com.algoExpert.demo.Repository.MemberRepository;
 import com.algoExpert.demo.Repository.ProjectRepository;
 import com.algoExpert.demo.Repository.TableRepository;
@@ -40,6 +42,9 @@ public class ProjectServiceImpl implements ProjectService {
     @Autowired
     private MemberService memberService;
 
+    @Autowired
+    ProjectMapper projectMapper;
+
     //  create project
     @Override
     public Integer createProject(Project project, int user_id) throws InvalidArgument {
@@ -52,7 +57,7 @@ public class ProjectServiceImpl implements ProjectService {
         Project savedProject = projectRepository.save(project);
 
         // save member
-        Member newMember = memberService.inviteMember(savedProject.getProject_id(),user.getUser_id());
+        Member newMember = memberService.inviteMember(savedProject.getProject_id(), user.getUser_id());
 
         // Create a default table using new member id
         tableService.createTable(savedProject.getProject_id(), newMember.getMember_id());
@@ -63,17 +68,22 @@ public class ProjectServiceImpl implements ProjectService {
         return savedProject.getProject_id();
     }
 
-    /*public Integer createProject(Project project, int user_id) throws InvalidArgument {
+/*
+* public Integer createProject(Project project, int user_id) throws InvalidArgument {
 //        find user by id
         User user = userRepository.findById(user_id).orElseThrow(()->new InvalidArgument("User with ID " + user_id + " not found"));
+//        Project project = projectMapper.projectDtoToProject(projectDto);
         project.setUser(user);
         Project savedProjects = projectRepository.save(project);
 
 //        add owner to the project as a member
-        List<Member> members = savedProjects.getMemberList();
+        List<Member> members = savedProjects.getMembersList();
+
+        System.err.println(555555);
         Member newMember = new Member(0, user.getUser_id(), savedProjects.getProject_id(), null);
         members.add(newMember);
-        project.setMemberList(members);
+        members.forEach(System.err::println);
+        project.setMembersList(members);
 
 //        create a default table
         tableService.createTable(project.getProject_id(), user.getUser_id());
@@ -84,21 +94,23 @@ public class ProjectServiceImpl implements ProjectService {
 
     //  get all projects
     @Override
-    public List<Project> getAllProjects() {
-        return projectRepository.findAll();
+    public List<ProjectDto> getAllProjects() {
+        List<Project> projects = projectRepository.findAll();
+        return projectMapper.projectDtos(projects);
     }
 
     //get one project
     @Override
-    public Project findProject(int project_id) {
-        return projectRepository.findById(project_id).get();
+    public ProjectDto findProject(int project_id) throws InvalidArgument {
+        Project project = projectRepository.findById(project_id).orElseThrow(() -> new InvalidArgument("Project with ID " + project_id + " not found"));
+        return projectMapper.projectToProjectDto(project);
     }
 
-//    delete project
+    //    delete project
     @Override
-    public List<Project> deleteProjectById(Integer projectId) throws InvalidArgument{
+    public List<Project> deleteProjectById(Integer projectId) throws InvalidArgument {
         Project project = projectRepository.findById(projectId)
-                    .orElseThrow(() -> new InvalidArgument("Project with ID " + projectId + " not found"));
+                .orElseThrow(() -> new InvalidArgument("Project with ID " + projectId + " not found"));
 
         // Delete associated tables
         for (Table table : project.getTables()) {
@@ -115,18 +127,32 @@ public class ProjectServiceImpl implements ProjectService {
         return projectRepository.findAll();
     }
 
-//    update project
+    //    update project
     @Override
-    public Project editProject(Project newProjectValue, int project_id) throws InvalidArgument{
-        return projectRepository.findById(project_id)
-                .map(existingProject -> {
-                    if (newProjectValue != null) {
-                        Optional.ofNullable(newProjectValue.getTitle()).ifPresent(existingProject::setTitle);
-                        Optional.ofNullable(newProjectValue.getDescription()).ifPresent(existingProject::setDescription);
-                    }
-                    return projectRepository.save(existingProject);
-                }).orElseThrow(() -> new InvalidArgument("Task with ID " + project_id + " not found"));
+    public ProjectDto editProject(ProjectDto projectDto, int project_id) throws InvalidArgument {
+
+        Project project = projectRepository.findById(project_id).map(existingProject -> {
+            if (projectDto != null) {
+                Optional.ofNullable(projectDto.getTitle()).ifPresent(existingProject::setTitle);
+                Optional.ofNullable(projectDto.getDescription()).ifPresent(existingProject::setDescription);
+            }
+            return projectRepository.save(existingProject);
+        }).orElseThrow(() -> new InvalidArgument("Task with ID " + project_id + " not found"));
+
+        return projectMapper.projectToProjectDto(project);
 
     }
+
+//    public Project editProject(Project newProjectValue, int project_id) throws InvalidArgument{
+//        return projectRepository.findById(project_id)
+//                .map(existingProject -> {
+//                    if (newProjectValue != null) {
+//                        Optional.ofNullable(newProjectValue.getTitle()).ifPresent(existingProject::setTitle);
+//                        Optional.ofNullable(newProjectValue.getDescription()).ifPresent(existingProject::setDescription);
+//                    }
+//                    return projectRepository.save(existingProject);
+//                }).orElseThrow(() -> new InvalidArgument("Task with ID " + project_id + " not found"));
+//
+//    }
 
 }
