+++ backend/src/main/java/com/algoExpert/demo/Service/Impl/TableServiceImpl.java
@@ -1,7 +1,11 @@
 package com.algoExpert.demo.Service.Impl;
 
+import com.algoExpert.demo.Dto.ProjectDto;
+import com.algoExpert.demo.Dto.TableDto;
 import com.algoExpert.demo.Entity.*;
 import com.algoExpert.demo.ExceptionHandler.InvalidArgument;
+import com.algoExpert.demo.Mapper.ProjectMapper;
+import com.algoExpert.demo.Mapper.TableMapper;
 import com.algoExpert.demo.Repository.MemberRepository;
 import com.algoExpert.demo.Repository.ProjectRepository;
 import com.algoExpert.demo.Repository.TableRepository;
@@ -30,8 +34,34 @@ public class TableServiceImpl implements TableService {
     @Autowired
     private MemberRepository memberRepository;
 
-//  create a new table
+    @Autowired
+    private TableMapper tableMapper;
+    @Autowired
+    ProjectMapper projectMapper;
+
+    //  create a new table
     @Override
+    public ProjectDto createTable(int project_id, int member_id) throws InvalidArgument {
+        Project project = projectRepository.findById(project_id).orElseThrow(() -> new InvalidArgument("Project with ID " + project_id + " not found"));
+
+        List<Table> tables = project.getTables();
+        int count = project.getTables().size();
+        Table table = new Table(0, "Table" + count, null);
+        Task task = new Task(0, "task", "description"
+                , member_id, "", "", "", "", null);
+
+
+        tables.add(table);
+        project.setTables(tables);
+        List<Task> taskList = new ArrayList<>();
+        taskList.add(task);
+        table.setTasks(taskList);
+
+        Project projectResult = projectRepository.save(project);
+        return projectMapper.projectToProjectDto(projectResult);
+    }
+
+    /*
     public Project createTable(int project_id, int member_id) throws InvalidArgument {
         // Retrieve the project by ID
         Project project = projectRepository.findById(project_id)
@@ -77,60 +107,43 @@ public class TableServiceImpl implements TableService {
             throw new InvalidArgument("Member ID " + member_id + " is not a member");
         }
     }
+     */
 
-
-    /*public Project createTable(int project_id, int member_id) throws InvalidArgument{
-        Project project =  projectRepository.findById(project_id).orElseThrow(()->
-        new InvalidArgument("Project with ID " + project_id + " not found"));
-
-        List<Table> tables = project.getTables();
-        int count = project.getTables().size();
-        Table table =new Table(0,"Table"+ count,null);
-        Task task=new Task(0,"task","description"
-                ,member_id,"","","","",null);
-
-
-        tables.add(table);
-        project.setTables(tables);
-        List<Task> taskList=new ArrayList<>();
-        taskList.add(task);
-        table.setTasks(taskList);
-
-        return projectRepository.save(project);
-    }*/
-
-//  get all table
+    //  get all table
     @Override
-    public List<Table> getAllTables() {
-        return tableRepository.findAll();
+    public List<TableDto> getAllTables() {
+        List<Table> tables = tableRepository.findAll();
+        return tableMapper.tableDtos(tables);
     }
 
-//  update table
+    //  update table
     @Override
-    public Table editTable(Table newTableValue,int table_id) throws InvalidArgument {
-        return tableRepository.findById(table_id)
-                .map(existingTable->{
-                    if(newTableValue !=null){
+    public TableDto editTable(TableDto newTableValue, int table_id) throws InvalidArgument {
+        Table table = tableRepository.findById(table_id)
+                .map(existingTable -> {
+                    if (newTableValue != null) {
                         Optional.ofNullable(newTableValue.getTable_name()).ifPresent(existingTable::setTable_name);
                     }
                     return tableRepository.save(existingTable);
                 }).orElseThrow(() -> new InvalidArgument("Task with ID " + table_id + " not found"));
+
+        return tableMapper.tableToTableDto(table);
 
     }
 
     //    delete table
     @Override
     @Transactional
-    public List<Table> deleteTable(Integer project_id, Integer table_id) throws InvalidArgument{
-        Project project = projectRepository.findById(project_id).orElseThrow(()->new InvalidArgument("Project with ID " + project_id + " not found"));
-        Table table = tableRepository.findById(table_id).orElseThrow(()->new InvalidArgument("Table with ID " + table_id + " not found"));
+    public List<TableDto> deleteTable(Integer project_id, Integer table_id) throws InvalidArgument {
+        Project project = projectRepository.findById(project_id).orElseThrow(() -> new InvalidArgument("Project with ID " + project_id + " not found"));
+        Table table = tableRepository.findById(table_id).orElseThrow(() -> new InvalidArgument("Table with ID " + table_id + " not found"));
 
         List<Table> tablesList = project.getTables();
         tablesList.remove(table);
         project.setTables(tablesList);
-
         projectRepository.save(project);
-        return tableRepository.findAll();
+        List<Table> tables = tableRepository.findAll();
+        return tableMapper.tableDtos(tables);
     }
 
 }
