+++ backend/src/main/java/com/algoExpert/demo/Service/Impl/TaskServiceImpl.java
@@ -1,7 +1,9 @@
 package com.algoExpert.demo.Service.Impl;
 
+import com.algoExpert.demo.Dto.TaskDto;
 import com.algoExpert.demo.Entity.*;
 import com.algoExpert.demo.ExceptionHandler.InvalidArgument;
+import com.algoExpert.demo.Mapper.TaskMapper;
 import com.algoExpert.demo.Repository.*;
 import com.algoExpert.demo.Service.TaskService;
 import jakarta.persistence.EntityManager;
@@ -15,28 +17,46 @@ import java.util.Optional;
 
 @Service
 public class TaskServiceImpl implements TaskService {
-     @Autowired
-     private TaskRepository taskRepository;
-     @Autowired
-     private ProjectRepository projectRepository;
+    @Autowired
+    private TaskRepository taskRepository;
+    @Autowired
+    private ProjectRepository projectRepository;
 
-     @PersistenceContext
-     private EntityManager entityManager;
+    @PersistenceContext
+    private EntityManager entityManager;
 
-     @Autowired
-     private AssigneesRepository assigneesRepository;
+    @Autowired
+    private AssigneesRepository assigneesRepository;
 
-     @Autowired
-     private TableRepository tableRepository;
+    @Autowired
+    private TableRepository tableRepository;
 
-     @Autowired
-     private CommentRepository commentRepository;
+    @Autowired
+    private CommentRepository commentRepository;
 
     @Autowired
     private MemberRepository memberRepository;
 
-//    create new task
+    @Autowired
+    private TaskMapper taskMapper;
+
+    //    create new task
     @Override
+    public Table createTask(int member_id, int table_id) throws InvalidArgument {
+
+        Table table = tableRepository.findById(table_id).orElseThrow(() -> new InvalidArgument("Table with ID " + table_id + " not found"));
+
+        List<Task> taskList = table.getTasks();
+        int count = taskList.size() + 1;
+        Task task = new Task(0, "task" + count, ""
+                , member_id, "", "", "", "", null);
+
+        taskList.add(task);
+        table.setTasks(taskList);
+
+        return tableRepository.save(table);
+    }
+    /*
     public Table createTask(int member_id, int table_id) throws InvalidArgument {
 
         // check if table and member exist
@@ -54,66 +74,71 @@ public class TaskServiceImpl implements TaskService {
         table.setTasks(taskList);
 
         return tableRepository.save(table);
-    }
+    }*/
 
-//  delete task
+
+    //    get all tasks
     @Override
-    @Transactional
-    public Table deleteTaskById(Integer task_id, Integer table_id) throws InvalidArgument{
-        Task storedTask = taskRepository.findById(task_id).orElseThrow(()->
-                new InvalidArgument("Task with ID " + task_id + " not found"));
-        Table table = tableRepository.findById(table_id).orElseThrow(()->
-                new InvalidArgument("Table with ID " + table_id + " not found"));
-
-            List<Comment> comments = storedTask.getComments();
-            if (!comments.isEmpty()) {
-                for (Comment comment : comments) {
-                    commentRepository.delete(comment);
-                }
-            }
-           List<Task> taskList = table.getTasks();
-           taskList.remove(storedTask);
-           table.setTasks(taskList);
-           tableRepository.save(table);
-        return table;
+    public List<TaskDto> getAllTask() {
+        List<Task> tasks = taskRepository.findAll();
+        return taskMapper.taskDtos(tasks);
     }
 
-//    get all tasks
+    //    update task
     @Override
-    public List<Task> getAllTask(){
-        return taskRepository.findAll();
-    }
-
-//    update task
-    public Task editTask(Task newTask,int task_id) throws InvalidArgument{
-        return   taskRepository.findById(newTask.getTask_id())
-                .map(existingTask ->{
-                    if (newTask != null){
-//                        Optional.ofNullable(newTask.getTask_id()).ifPresent(existingTask::setTask_id);
-                        Optional.ofNullable(newTask.getTitle()).ifPresent(existingTask::setTitle);
-                        Optional.ofNullable(newTask.getDescription()).ifPresent(existingTask::setDescription);
-                        Optional.ofNullable(newTask.getStart_date()).ifPresent(existingTask::setStart_date);
-                        Optional.ofNullable(newTask.getEnd_date()).ifPresent(existingTask::setEnd_date);
-                        Optional.ofNullable(newTask.getStatus()).ifPresent(existingTask::setStatus);
-                        Optional.ofNullable(newTask.getPriority()).ifPresent(existingTask::setPriority);
+    public TaskDto editTask(TaskDto newTaskDto) throws InvalidArgument {
+        Task task = taskRepository.findById(newTaskDto.getTask_id())
+                .map(existingTask -> {
+                    if (newTaskDto != null) {
+                        Optional.ofNullable(newTaskDto.getTitle()).ifPresent(existingTask::setTitle);
+                        Optional.ofNullable(newTaskDto.getDescription()).ifPresent(existingTask::setDescription);
+                        Optional.ofNullable(newTaskDto.getStart_date()).ifPresent(existingTask::setStart_date);
+                        Optional.ofNullable(newTaskDto.getEnd_date()).ifPresent(existingTask::setEnd_date);
+                        Optional.ofNullable(newTaskDto.getStatus()).ifPresent(existingTask::setStatus);
+                        Optional.ofNullable(newTaskDto.getPriority()).ifPresent(existingTask::setPriority);
                     }
                     return taskRepository.save(existingTask);
-                }).orElseThrow(() -> new InvalidArgument("Task with ID " + task_id + " not found"));
+                }).orElseThrow(() -> new InvalidArgument("Task with ID " + newTaskDto.getTask_id() + " not found"));
+        return taskMapper.taskToTaskDto(task);
     }
 
     //duplicate task
     @Override
-    public Table duplicateTask(Task task, Integer table_id){
+    public Table duplicateTask(Task task, Integer table_id) {
         Table table = tableRepository.findById(table_id).get();
 
-        Task newTask = new Task(0,task.getTitle(),task.getDescription()
-                ,task.getOwner(),task.getStart_date(),task.getEnd_date(),task.getStatus(),
-                task.getPriority(),null);
-        List<Task> taskList =table.getTasks();
+        Task newTask = new Task(0, task.getTitle(), task.getDescription()
+
+                , task.getOwner(), task.getStart_date(), task.getEnd_date(), task.getStatus(),
+                task.getPriority(), null);
+        List<Task> taskList = table.getTasks();
+
 
         taskList.add(newTask);
         table.setTasks(taskList);
-        return  tableRepository.save(table);
+        return tableRepository.save(table);
+    }
+
+    //  delete task
+    @Override
+    @Transactional
+    public Table deleteTaskById(Integer task_id, Integer table_id) throws InvalidArgument {
+        Task storedTask = taskRepository.findById(task_id).orElseThrow(() ->
+                new InvalidArgument("Task with ID " + task_id + " not found"));
+        Table table = tableRepository.findById(table_id).orElseThrow(() ->
+                new InvalidArgument("Table with ID " + table_id + " not found"));
+
+        List<Comment> comments = storedTask.getComments();
+        if (!comments.isEmpty()) {
+            for (Comment comment : comments) {
+                commentRepository.delete(comment);
+            }
+        }
+        List<Task> taskList = table.getTasks();
+        taskList.remove(storedTask);
+        table.setTasks(taskList);
+        tableRepository.save(table);
+        return table;
     }
 
 }
