+++ backend/src/main/java/com/algoExpert/demo/Service/TaskService.java
@@ -15,77 +15,24 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class TaskService {
-    @Autowired
-    private TaskRepository taskRepository;
-    @Autowired
-    private TableRepository tableRepository;
-    @Autowired
-    private CommentRepository commentRepository;
-    @Autowired
-    private TaskMapper taskMapper;
-
-    public interface TaskService {
+public interface TaskService {
 
 
-        //    create new task
-        public Table createTask(int member_id, int table_id) throws InvalidArgument {
+    //    create new task
+    public Table createTask(int member_id, int table_id) throws InvalidArgument;
 
-            Table table = tableRepository.findById(table_id).orElseThrow(() -> new InvalidArgument("Table with ID " + table_id + " not found"));
+    //    get all tasks
+    public List<TaskDto> getAllTask();
 
-            List<Task> taskList = table.getTasks();
-            int count = taskList.size() + 1;
-            Task task = new Task(0, "task" + count, ""
-                    , member_id, "", "", "", "", null);
+    //    update task
+    public TaskDto editTask(TaskDto newTaskDto) throws InvalidArgument;
 
-            taskList.add(task);
-            table.setTasks(taskList);
+    //duplicate task
+    public Table duplicateTask(Task task, Integer table_id);
 
-            return tableRepository.save(table);
-        }
+    //  delete task
+    @Transactional
+    Table deleteTaskById(Integer taskId, Integer table_id) throws InvalidArgument;
 
-        //  delete task
-        @Transactional
-        Table deleteTaskById(Integer taskId, Integer table_id) throws InvalidArgument;
-
-        //    get all tasks
-        public List<TaskDto> getAllTask() {
-            List<Task> tasks = taskRepository.findAll();
-            return taskMapper.taskDtos(tasks);
-        }
-
-        //    update task
-        public TaskDto editTask(TaskDto newTaskDto) throws InvalidArgument {
-            Task task = taskRepository.findById(newTaskDto.getTask_id())
-                    .map(existingTask -> {
-                        if (newTaskDto != null) {
-                            Optional.ofNullable(newTaskDto.getTitle()).ifPresent(existingTask::setTitle);
-                            Optional.ofNullable(newTaskDto.getDescription()).ifPresent(existingTask::setDescription);
-                            Optional.ofNullable(newTaskDto.getStart_date()).ifPresent(existingTask::setStart_date);
-                            Optional.ofNullable(newTaskDto.getEnd_date()).ifPresent(existingTask::setEnd_date);
-                            Optional.ofNullable(newTaskDto.getStatus()).ifPresent(existingTask::setStatus);
-                            Optional.ofNullable(newTaskDto.getPriority()).ifPresent(existingTask::setPriority);
-                        }
-                        return taskRepository.save(existingTask);
-                    }).orElseThrow(() -> new InvalidArgument("Task with ID " + newTaskDto.getTask_id() + " not found"));
-            return taskMapper.taskToTaskDto(task);
-        }
-
-        //duplicate task
-        public Table duplicateTask(Task task, Integer table_id) {
-            Table table = tableRepository.findById(table_id).get();
-
-            Task newTask = new Task(0, task.getTitle(), task.getDescription()
-
-                    , task.getOwner(), task.getStart_date(), task.getEnd_date(), task.getStatus(),
-                    task.getPriority(), null);
-            List<Task> taskList = table.getTasks();
-
-
-            taskList.add(newTask);
-            table.setTasks(taskList);
-            return tableRepository.save(table);
-        }
-
-    }
+}
