+++ backend/src/main/java/com/algoExpert/demo/Service/UserService.java
@@ -15,45 +15,15 @@ import java.util.stream.Collectors;
 @Service
 public interface UserService {
 
-    @Autowired
-    UserMapper userMapper;
+    //    create user
+    UserDto create(UserDto userDto);
 
-//    create user
-    public UserDto create(UserDto userDto){
-        User user = userMapper.userDtoToUser(userDto);
-        User userResults = userRepository.save(user);
-        return userMapper.userToUserDto(userResults);
-    }
+    // get all users
+    List<UserDto> getUsers();
 
-// get all users
-    public List<UserDto> getUsers(){
-        List<User> users = userRepository.findAll();
-        return userMapper.usersToUserDtos(users);
-    }
+    //    delete user by id
+    List<UserDto> deleteUser(int userId)
 
-//    delete user by id
-    public List<UserDto> deleteUser(int userId){
-        userRepository.deleteById(userId);
-        return userRepository.findAll()
-                .stream().map(user -> new UserDto(user.getUser_id(),
-                        user.getUsername(),
-                        user.getEmail())).collect(Collectors.toList());
-    }
-
-
-    public List<Project> getUserProjectIds(int userId) {
-        // Find all members
-        List<Member> memberList = memberRepository.findAll();
-
-        // Filter members by user_id and map them to project ids
-        List<Integer> userProjectIds = memberList.stream()
-                .filter(member -> member.getUser_id() == userId)
-                .map(Member::getProject_id) // Assuming you have a method getProject_id() in Member class
-                .collect(Collectors.toList());
-
-        return projectRepository.findAllById(userProjectIds);
-    }
-
-
-
+    //    get User Project Ids
+    List<Project> getUserProjectIds(int userId);
 }
